+++ CHANGELOG.md
@@ -6,7 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-* Improved test coverage for `McAvoy_Query` class.
+* Improved test coverage for `McAvoy_Query` class and `McAvoy\Admin\maybe_delete_queries()`.
+* Added additional documentation relating to the current stable tag and test coverage.
 
 
 ## [0.1.2] - 2016-03-29